# docs/swagger.yaml
basePath: /
definitions:
  app.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      errMsg:
        type: string
      msg:
        type: string
    type: object
  request.ReqAddUserForm:
    properties:
      address:
        description: 地址
        type: string
      nick_name:
        description: 姓名
        type: string
      password:
        description: 登录密码
        type: string
      phone_number:
        description: 电话号码
        type: string
      user_name:
        description: 用户名
        type: string
    required:
    - password
    - user_name
    type: object
  request.ReqCrawler:
    properties:
      type:
        type: string
      uri:
        type: string
    required:
    - type
    - uri
    type: object
  request.ReqLoginForm:
    properties:
      password:
        type: string
      user_name:
        description: 页面大小
        type: string
    type: object
  request.ReqUpdateUserForm:
    properties:
      address:
        description: 地址
        type: string
      id:
        description: 主键
        type: integer
      nick_name:
        description: 姓名
        type: string
      password:
        description: 登录密码
        type: string
      phone_number:
        description: 电话号码
        type: string
      status:
        description: 账户状态（1：暂停，2：启用）
        type: integer
      user_name:
        description: 用户名
        type: string
    required:
    - id
    type: object
  request.ReqUserIdForm:
    properties:
      id:
        description: 页面大小
        type: integer
    required:
    - id
    type: object
info:
  contact: {}
  description: An example of gin
  license:
    name: MIT
    url: https://Pay/blob/master/LICENSE
  termsOfService: https://github.com/EDDYCJY/go-gin-example
  title: Golang Gin API
  version: "1.0"
paths:
  /api/v1/article/crawler:
    post:
      consumes:
      - application/json
      description: 文章爬取
      parameters:
      - description: reqBody
        in: body
        name: form
        required: true
        schema:
          $ref: '#/definitions/request.ReqCrawler'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 文章爬取
      tags:
      - 文章
  /api/v1/article/list:
    get:
      consumes:
      - application/json
      description: 文章查询
      parameters:
      - description: 当前页面
        in: query
        name: page
        type: integer
      - description: 页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 文章列表
      tags:
      - 文章
  /api/v1/article/weibo:
    get:
      consumes:
      - application/json
      description: 文章查询
      parameters:
      - description: 当前页面
        in: query
        name: page
        type: integer
      - description: 页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 文章列表
      tags:
      - 文章
  /api/v1/general/weibo/list:
    get:
      consumes:
      - application/json
      description: 获取微博热点数据
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 获取微博热点数据
      tags:
      - 通用
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: 登录
      parameters:
      - description: 登录
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.ReqLoginForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 登录
      tags:
      - 用户
  /api/v1/register:
    post:
      consumes:
      - application/json
      description: 添加用户信息
      parameters:
      - description: reqBody
        in: body
        name: form
        required: true
        schema:
          $ref: '#/definitions/request.ReqAddUserForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 注册
      tags:
      - 用户
  /api/v1/system/initScript:
    get:
      consumes:
      - application/json
      description: 初始化数据库
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 初始化数据库
      tags:
      - 系统设置
  /api/v1/user/delete:
    post:
      consumes:
      - application/json
      description: 删除用户信息
      parameters:
      - description: Id
        in: body
        name: form
        required: true
        schema:
          $ref: '#/definitions/request.ReqUserIdForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - Token: []
      summary: 删除用户
      tags:
      - 用户
  /api/v1/user/get:
    get:
      consumes:
      - application/json
      description: 获取用户信息
      parameters:
      - description: 用户ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - Token: []
      summary: 获取单个用户信息
      tags:
      - 用户
  /api/v1/user/info:
    get:
      consumes:
      - application/json
      description: 获取用户信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - Token: []
      summary: 获取当前用户信息
      tags:
      - 用户
  /api/v1/user/list:
    get:
      consumes:
      - application/json
      description: 获取用户信息
      parameters:
      - description: 当前页面
        in: query
        name: page
        type: integer
      - description: 页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - Token: []
      summary: 获取用户信息
      tags:
      - 用户
  /api/v1/user/logout:
    post:
      consumes:
      - application/json
      description: 登出
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - Token: []
      summary: 登出
      tags:
      - 用户
  /api/v1/user/update:
    post:
      consumes:
      - application/json
      description: 修改用户信息
      parameters:
      - description: reqBody
        in: body
        name: form
        required: true
        schema:
          $ref: '#/definitions/request.ReqUpdateUserForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - Token: []
      summary: 修改用户
      tags:
      - 用户
  /api/v1/ws/send:
    get:
      consumes:
      - application/json
      description: 发送消息
      parameters:
      - description: 当前页面
        in: query
        name: page
        type: integer
      - description: 页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 发送消息
      tags:
      - websocket Test
securityDefinitions:
  Token:
    in: header
    name: token
    type: apiKey
swagger: "2.0"